update latest ruby client (#182)
### What changed? Why?
follow up change to this PR
https://github.com/coinbase/coinbase-sdk-ruby/pull/181

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
