+++ lib/coinbase/client.rb
@@ -43,6 +43,8 @@ Coinbase::Client.autoload :CreateTransferRequest, 'coinbase/client/models/create
 Coinbase::Client.autoload :CreateWalletRequest, 'coinbase/client/models/create_wallet_request'
 Coinbase::Client.autoload :CreateWalletRequestWallet, 'coinbase/client/models/create_wallet_request_wallet'
 Coinbase::Client.autoload :CreateWebhookRequest, 'coinbase/client/models/create_webhook_request'
+Coinbase::Client.autoload :ERC20TransferEvent, 'coinbase/client/models/erc20_transfer_event'
+Coinbase::Client.autoload :ERC721TransferEvent, 'coinbase/client/models/erc721_transfer_event'
 Coinbase::Client.autoload :Error, 'coinbase/client/models/error'
 Coinbase::Client.autoload :EthereumTransaction, 'coinbase/client/models/ethereum_transaction'
 Coinbase::Client.autoload :EthereumTransactionAccess, 'coinbase/client/models/ethereum_transaction_access'
@@ -102,6 +104,7 @@ Coinbase::Client.autoload :WebhookList, 'coinbase/client/models/webhook_list'
 # APIs
 Coinbase::Client.autoload :AddressesApi, 'coinbase/client/api/addresses_api'
 Coinbase::Client.autoload :AssetsApi, 'coinbase/client/api/assets_api'
+Coinbase::Client.autoload :BalanceHistoryApi, 'coinbase/client/api/balance_history_api'
 Coinbase::Client.autoload :ContractEventsApi, 'coinbase/client/api/contract_events_api'
 Coinbase::Client.autoload :ContractInvocationsApi, 'coinbase/client/api/contract_invocations_api'
 Coinbase::Client.autoload :ExternalAddressesApi, 'coinbase/client/api/external_addresses_api'
