+++ lib/coinbase/client/api/external_addresses_api.rb
@@ -94,91 +94,6 @@ module Coinbase::Client
       return data, status_code, headers
     end
 
-    # Get address balance history for asset
-    # List the historical balance of an asset in a specific address.
-    # @param network_id [String] The ID of the blockchain network
-    # @param address_id [String] The ID of the address to fetch the historical balance for.
-    # @param asset_id [String] The symbol of the asset to fetch the historical balance for.
-    # @param [Hash] opts the optional parameters
-    # @option opts [Integer] :limit A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10.
-    # @option opts [String] :page A cursor for pagination across multiple pages of results. Don&#39;t include this parameter on the first call. Use the next_page value returned in a previous response to request subsequent results.
-    # @return [AddressHistoricalBalanceList]
-    def list_address_historical_balance(network_id, address_id, asset_id, opts = {})
-      data, _status_code, _headers = list_address_historical_balance_with_http_info(network_id, address_id, asset_id, opts)
-      data
-    end
-
-    # Get address balance history for asset
-    # List the historical balance of an asset in a specific address.
-    # @param network_id [String] The ID of the blockchain network
-    # @param address_id [String] The ID of the address to fetch the historical balance for.
-    # @param asset_id [String] The symbol of the asset to fetch the historical balance for.
-    # @param [Hash] opts the optional parameters
-    # @option opts [Integer] :limit A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10.
-    # @option opts [String] :page A cursor for pagination across multiple pages of results. Don&#39;t include this parameter on the first call. Use the next_page value returned in a previous response to request subsequent results.
-    # @return [Array<(AddressHistoricalBalanceList, Integer, Hash)>] AddressHistoricalBalanceList data, response status code and response headers
-    def list_address_historical_balance_with_http_info(network_id, address_id, asset_id, opts = {})
-      if @api_client.config.debugging
-        @api_client.config.logger.debug 'Calling API: ExternalAddressesApi.list_address_historical_balance ...'
-      end
-      # verify the required parameter 'network_id' is set
-      if @api_client.config.client_side_validation && network_id.nil?
-        fail ArgumentError, "Missing the required parameter 'network_id' when calling ExternalAddressesApi.list_address_historical_balance"
-      end
-      # verify the required parameter 'address_id' is set
-      if @api_client.config.client_side_validation && address_id.nil?
-        fail ArgumentError, "Missing the required parameter 'address_id' when calling ExternalAddressesApi.list_address_historical_balance"
-      end
-      # verify the required parameter 'asset_id' is set
-      if @api_client.config.client_side_validation && asset_id.nil?
-        fail ArgumentError, "Missing the required parameter 'asset_id' when calling ExternalAddressesApi.list_address_historical_balance"
-      end
-      if @api_client.config.client_side_validation && !opts[:'page'].nil? && opts[:'page'].to_s.length > 5000
-        fail ArgumentError, 'invalid value for "opts[:"page"]" when calling ExternalAddressesApi.list_address_historical_balance, the character length must be smaller than or equal to 5000.'
-      end
-
-      # resource path
-      local_var_path = '/v1/networks/{network_id}/addresses/{address_id}/balance_history/{asset_id}'.sub('{' + 'network_id' + '}', CGI.escape(network_id.to_s)).sub('{' + 'address_id' + '}', CGI.escape(address_id.to_s)).sub('{' + 'asset_id' + '}', CGI.escape(asset_id.to_s))
-
-      # query parameters
-      query_params = opts[:query_params] || {}
-      query_params[:'limit'] = opts[:'limit'] if !opts[:'limit'].nil?
-      query_params[:'page'] = opts[:'page'] if !opts[:'page'].nil?
-
-      # header parameters
-      header_params = opts[:header_params] || {}
-      # HTTP header 'Accept' (if needed)
-      header_params['Accept'] = @api_client.select_header_accept(['application/json'])
-
-      # form parameters
-      form_params = opts[:form_params] || {}
-
-      # http body (model)
-      post_body = opts[:debug_body]
-
-      # return_type
-      return_type = opts[:debug_return_type] || 'AddressHistoricalBalanceList'
-
-      # auth_names
-      auth_names = opts[:debug_auth_names] || []
-
-      new_options = opts.merge(
-        :operation => :"ExternalAddressesApi.list_address_historical_balance",
-        :header_params => header_params,
-        :query_params => query_params,
-        :form_params => form_params,
-        :body => post_body,
-        :auth_names => auth_names,
-        :return_type => return_type
-      )
-
-      data, status_code, headers = @api_client.call_api(:GET, local_var_path, new_options)
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "API called: ExternalAddressesApi#list_address_historical_balance\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
-      end
-      return data, status_code, headers
-    end
-
     # List transactions for an address.
     # List all transactions that interact with the address.
     # @param network_id [String] The ID of the blockchain network
